# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deep Analysis Financial Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stocks/screen:
    get:
      operationId: screenStocks
      tags: [stocks]
      summary: Screen high-growth infrastructure stocks
      description: Returns stocks matching the LITE-style breakout criteria
      parameters:
        - name: riskProfile
          in: query
          required: false
          schema:
            type: string
            enum: [high_growth, balanced, value_stability]
          description: Risk/return profile filter
        - name: sector
          in: query
          required: false
          schema:
            type: string
          description: Sector filter
      responses:
        "200":
          description: Screened stocks matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScreenedStock"
  /stocks/{symbol}/quote:
    get:
      operationId: getStockQuote
      tags: [stocks]
      summary: Get real-time stock quote and fundamentals
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock quote and fundamentals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockQuote"
        "404":
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /stocks/{symbol}/news:
    get:
      operationId: getStockNews
      tags: [stocks]
      summary: Get latest news for a stock
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Latest news articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ScreenedStock:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        sector:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        revenueGrowth:
          type: number
        peRatio:
          type: number
        marketCap:
          type: number
        institutionalOwnership:
          type: number
        macroKeywords:
          type: array
          items:
            type: string
        geopoliticalScore:
          type: number
          description: 0-100 geopolitical exposure score
        breakoutScore:
          type: number
          description: 0-100 LITE-style breakout score
        riskProfile:
          type: string
          enum: [high_growth, balanced, value_stability]
        isSoldOut:
          type: boolean
      required:
        - symbol
        - name
        - sector
        - price
        - change
        - changePercent
        - revenueGrowth
        - peRatio
        - marketCap
        - institutionalOwnership
        - macroKeywords
        - geopoliticalScore
        - breakoutScore
        - riskProfile
        - isSoldOut
    StockQuote:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        price:
          type: number
        open:
          type: number
        high:
          type: number
        low:
          type: number
        volume:
          type: number
        peRatio:
          type: number
        eps:
          type: number
        revenueGrowth:
          type: number
        grossMargin:
          type: number
        institutionalOwnership:
          type: number
        fiftyTwoWeekHigh:
          type: number
        fiftyTwoWeekLow:
          type: number
        marketCap:
          type: number
        dividendYield:
          type: number
      required:
        - symbol
        - name
        - price
        - open
        - high
        - low
        - volume
        - peRatio
        - eps
        - revenueGrowth
        - grossMargin
        - institutionalOwnership
        - fiftyTwoWeekHigh
        - fiftyTwoWeekLow
        - marketCap
        - dividendYield
    NewsArticle:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        url:
          type: string
        source:
          type: string
        publishedAt:
          type: string
        sentiment:
          type: string
          enum: [Bullish, Bearish, Neutral, Somewhat-Bullish, Somewhat-Bearish]
        sentimentScore:
          type: number
        relevanceScore:
          type: number
      required:
        - title
        - summary
        - url
        - source
        - publishedAt
        - sentiment
        - sentimentScore
        - relevanceScore
...