# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Analysis Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { NewsArticleSentiment } from "./newsArticleSentiment";

export interface NewsArticle {
  title: string;
  summary: string;
  url: string;
  source: string;
  publishedAt: string;
  sentiment: NewsArticleSentiment;
  sentimentScore: number;
  relevanceScore: number;
}