# lib/api-zod/src/generated/types/screenedStock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deep Analysis Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ScreenedStockRiskProfile } from "./screenedStockRiskProfile";

export interface ScreenedStock {
  symbol: string;
  name: string;
  sector: string;
  price: number;
  change: number;
  changePercent: number;
  revenueGrowth: number;
  peRatio: number;
  marketCap: number;
  institutionalOwnership: number;
  macroKeywords: string[];
  /** 0-100 geopolitical exposure score */
  geopoliticalScore: number;
  /** 0-100 LITE-style breakout score */
  breakoutScore: number;
  riskProfile: ScreenedStockRiskProfile;
  isSoldOut: boolean;
}